+++ src/Reader.php
@@ -17,6 +17,7 @@ use CallbackFilterIterator;
 use Closure;
 use Iterator;
 use JsonSerializable;
+use League\Csv\Fragment\Expression;
 use League\Csv\Serializer\Denormalizer;
 use League\Csv\Serializer\MappingFailed;
 use League\Csv\Serializer\TypeCastingFailed;
@@ -413,12 +414,12 @@ class Reader extends AbstractCsv implements TabularDataReader, JsonSerializable
         return Statement::create()->orderBy($orderBy)->process($this);
     }
 
-    public function matching(string $expression): iterable
+    public function matching(Expression|string $expression): iterable
     {
-        return FragmentFinder::create()->findAll($expression, $this);
+        return Expression::from($expression)->fragment($this);
     }
 
-    public function matchingFirst(string $expression): ?TabularDataReader
+    public function matchingFirst(Expression|string $expression): ?TabularDataReader
     {
         return FragmentFinder::create()->findFirst($expression, $this);
     }
@@ -427,7 +428,7 @@ class Reader extends AbstractCsv implements TabularDataReader, JsonSerializable
      * @throws SyntaxError
      * @throws FragmentNotFound
      */
-    public function matchingFirstOrFail(string $expression): TabularDataReader
+    public function matchingFirstOrFail(Expression|string $expression): TabularDataReader
     {
         return FragmentFinder::create()->findFirstOrFail($expression, $this);
     }